requirements.txt
# Core packages
requests==2.31.0
numpy>=1.26.0
pandas>=2.1.0
six>=1.16.0

# Development and Testing
python-dotenv==1.0.0
pytest==7.4.2

# Database
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9

# Kafka
kafka-python-ng==2.2.3

# Spark
pyspark==3.5.0

# The following packages will be installed later after basic setup
# apache-airflow==2.7.1
# black==23.9.1
# flake8==6.1.0 